feat: directory branched, functions verified
