+++ src/utils/wallet/index.js
@@ -1,0 +1,2 @@
+import core from "./core.js"
+export default core
