--- src/utils/walletLibrary/index.js
@@ -1,6 +1,0 @@
-const library=require("./library")
-const run= async ()=> {
-console.log(library.createWallet())
- console.log(library.loadWallet);
-}
-run()
